slight clean up and organization
